requirements.txt
# Requisitos generales
requests==2.28.2
SQLAlchemy==1.4.47
psycopg2-binary==2.9.5
python-dotenv==1.0.0
kafka-python==2.0.2
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.2.2
pytest==7.3.1
pytest-cov==4.1.0
black==23.3.0
flake8==6.0.0
mypy==1.3.0
colorama==0.4.6

# FastAPI y servicios web
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7

# ML y IA
river==0.14.0
joblib==1.2.0
spacy==3.5.2
faiss-cpu==1.7.4
transformers==4.28.1
torch==2.0.1
nltk==3.8.1

# Monitoreo
prometheus-client==0.16.0

# Herramientas de resiliencia
pyrate-limiter==2.10.0
pybreaker==1.0.2
